chore(reverb): setting reverb variables

+++ docker/bootstrap.sh
@@ -18,10 +18,10 @@ if [ -d bootstrap ]; then
   chmod -R 755 bootstrap
 fi
 
-if [ "$env" != "production" ]; then
+if [ "$env" != "production"  ] && [ "$role" = "app" ]; then
   php artisan migrate:fresh --seed
 
-elif [ "$env" = "local" ]; then
+elif [ "$env" = "local" ] && [ "$role" = "app" ]; then
   if [ ! -f .env ]; then
     cp .env.example .env
   fi
@@ -30,11 +30,14 @@ else
 fi
 
 
+if [ "$role" = "app" ]; then
 
   php artisan cache:clear
   php artisan config:clear
   php artisan route:clear
   php artisan view:clear
+
+fi
